requirements.txt
click
Flask
Flask-CORS
simplejson
pyzmq
greenlet>=0.4.14
gevent>=1.3.5
PyYaml
requests
minio
pip
schematics
git+https://github.com/iterait/emloop.git
git+https://github.com/iterait/apistrap.git

# testing
pytest
pytest-mock
pytest-flask
pytest-forked
